refactor: 에러코드 타입 이름 변경
