+++ src/main/java/roomescape/exception/ErrorCodeWithHttpStatusCode.java
@@ -1,0 +1,9 @@
+package roomescape.exception;
+
+import org.springframework.http.HttpStatusCode;
+
+public interface ErrorCodeWithHttpStatusCode {
+    HttpStatusCode httpStatusCode();
+
+    String message();
+}
